novo/requirements.txt
google-api-python-client>=2.105.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
google-generativeai>=0.3.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0